adding wait for removal logic
this change add some logic to detect when machines are deleting but
replacements are ready. it also fixes the tests with respect to making
sure that deleted machines are not ready.

it is worth noting that this change moves where the waitingForRemoval
log message is emitted to happen in the waitForPendingMachines function.
as such some of the tests have been adjusted to not expect the
waitingForRemoval log as it is secondary to the main action.